+++ mRemoteNG/Messages/ReconnectEvent.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace mRemoteNG.Messages
+{
+    public static class ReconnectEvent
+    {
+        public delegate void EventHandler(object sender, EventArgs e);
+    }
+}